requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.10
Django==4.0.1
django-environ==0.8.1
django-jalali-date==1.0.1
gevent==21.12.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
jdatetime==3.8.1
Pillow==9.0.0
psycopg2-binary==2.9.3
requests==2.27.1
sqlparse==0.4.2
urllib3==1.26.8
uWSGI==2.0.20
zope.event==4.5.0
zope.interface==5.4.0
